Credit support

+++ src/main/java/qiwi/controller/BankController.java
@@ -6,8 +6,10 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import qiwi.dao.impl.BankDAO;
 import qiwi.dao.impl.ClientDAO;
+import qiwi.dao.impl.CreditDAO;
 import qiwi.model.Bank;
 import qiwi.model.Client;
+import qiwi.model.Credit;
 import qiwi.model.input.BankInput;
 import qiwi.model.input.ClientInput;
 import qiwi.model.input.CreditInput;
@@ -22,25 +24,28 @@ public class BankController {
     private BankDAO bankDAO;
     @Autowired
     private ClientDAO clientDAO;
+    @Autowired
+    private CreditDAO creditDAO;
 
-    private void setUpView(Model model, ClientInput clientInput, BankInput bankInput) {
+    private void setUpView(Model model, ClientInput clientInput, CreditInput creditInput, BankInput bankInput) {
         model.addAttribute("banks", bankDAO.findAll());
         model.addAttribute("clientInput", clientInput);
+        model.addAttribute("creditInput", creditInput);
         model.addAttribute("bankInput", bankInput);
     }
 
     @PostMapping("/add")
     public String add(@ModelAttribute BankInput input, Model model) {
         if (input.hasEmptyFields()) {
-            setUpView(model, new ClientInput(), input);
-            model.addAttribute("emptyFieldsMessage", "");
+            setUpView(model, new ClientInput(), new CreditInput(), input);
+            model.addAttribute("emptyFieldsBankMessage", "");
             return "banks";
         }
 
         Bank bank = new Bank(input);
         if (bankDAO.exists(bank)) {
-            setUpView(model, new ClientInput(), input);
-            model.addAttribute("alreadyExistsMessage", "");
+            setUpView(model, new ClientInput(), new CreditInput(), input);
+            model.addAttribute("alreadyExistsBankMessage", "");
             return "banks";
         }
 
@@ -52,14 +57,14 @@ public class BankController {
     @PostMapping("/edit/{id}")
     public String edit(@ModelAttribute BankInput input, Model model) {
         if (input.hasEmptyFields()) {
-            setUpView(model, new ClientInput(), input);
+            setUpView(model, new ClientInput(), new CreditInput(), input);
             model.addAttribute("emptyFieldsBankMessage", "");
             return "banks";
         }
 
         Bank bank = bankDAO.getBankByName(input.getName());
         if (bankDAO.getBankByName(input.getNewName()) != null) {
-            setUpView(model, new ClientInput(), input);
+            setUpView(model, new ClientInput(), new CreditInput(), input);
             model.addAttribute("alreadyExistsBankNameMessage", "");
             return "banks";
         }
@@ -79,7 +84,7 @@ public class BankController {
                 if (!input.getPassport().isEmpty()) {
                     if (bankDAO.exists(input.getBank(), client)) {
                         model.addAttribute("alreadyExistsMessage", "");
-                        setUpView(model, input, new BankInput());
+                        setUpView(model, input, new CreditInput(), new BankInput());
                         return "banks";
                     } else {
                         client = clientDAO.getClientByPassport(client.getPassport());
@@ -93,12 +98,12 @@ public class BankController {
             }
 
             model.addAttribute("emptyFieldsMessage", "");
-            setUpView(model, input, new BankInput());
+            setUpView(model, input, new CreditInput(), new BankInput());
             return "banks";
         }
 
-        if (!Validator.isValid(input)) {
-            setUpView(model, input, new BankInput());
+        if (!Validator.Client.isValid(input)) {
+            setUpView(model, input, new CreditInput(), new BankInput());
             model.addAttribute("invalidFieldsMessage", "");
             return "banks";
         }
@@ -106,7 +111,7 @@ public class BankController {
         Client client = new Client(input);
         if (bankDAO.exists(input.getBank(), client)) {
             model.addAttribute("alreadyExistsMessage", "");
-            setUpView(model, input, new BankInput());
+            setUpView(model, input, new CreditInput(), new BankInput());
             return "banks";
         }
 
@@ -123,7 +128,51 @@ public class BankController {
 
     @PostMapping("/addCredit")
     public String addCredit(@ModelAttribute CreditInput input, Model model) {
+        if (input.hasEmptyFields()) {
+            Credit credit = new Credit(input);
 
+            if (creditDAO.exists(credit.getId())) {
+                if (!input.getId().isEmpty()) {
+                    if (bankDAO.exists(input.getBank(), credit)) {
+                        model.addAttribute("alreadyExistsMessage", "");
+                        setUpView(model, new ClientInput(), input, new BankInput());
+                        return "banks";
+                    } else {
+                        credit = creditDAO.getCreditById(credit.getId());
+                        Bank bank = bankDAO.getBankByName(input.getBank());
+                        bank.addCredit(credit);
+                        bankDAO.add(bank);
+
+                        return "redirect:/banks/";
+                    }
+                }
+            }
+
+            model.addAttribute("emptyFieldsMessage", "");
+            setUpView(model, new ClientInput(), input, new BankInput());
+            return "banks";
+        }
+
+        if (!Validator.Credit.isValid(input)) {
+            setUpView(model, new ClientInput(), input, new BankInput());
+            model.addAttribute("invalidFieldsMessage", "");
+            return "banks";
+        }
+
+        Credit credit = new Credit(input);
+        if (bankDAO.exists(input.getBank(), credit)) {
+            model.addAttribute("alreadyExistsMessage", "");
+            setUpView(model, new ClientInput(), input, new BankInput());
+            return "banks";
+        }
+
+        Bank bank = bankDAO.getBankByName(input.getBank());
+        bank.addCredit(credit);
+        bankDAO.add(bank);
+
+        if (!creditDAO.exists(credit.getId())) {
+            creditDAO.add(credit);
+        }
 
         return "redirect:/banks/";
     }
@@ -135,20 +184,26 @@ public class BankController {
     }
 
     @GetMapping("/deleteClient/{id}")
-    public String deleteClient(@PathVariable UUID id) {
+    public String deleteClient(@PathVariable UUID id, Model model) {
+        if (clientDAO.hasCredit(id)) {
+            setUpView(model, new ClientInput(), new CreditInput(), new BankInput());
+            model.addAttribute("hasCreditBankMessage", "");
+            return "banks";
+        }
+
         bankDAO.deleteClientById(id);
         return "redirect:/banks/";
     }
 
     @GetMapping("/deleteCredit/{id}")
     public String deleteCredit(@PathVariable UUID id) {
-//        bankDAO.deleteCreditById(id);
+        creditDAO.delete(id);
         return "redirect:/banks/";
     }
 
     @GetMapping("/")
     public String showAllBanks(Model model) {
-        setUpView(model, new ClientInput(), new BankInput());
+        setUpView(model, new ClientInput(), new CreditInput(), new BankInput());
         return "banks";
     }
 }
